moved SavedView related code to extras
