+++ nautobot/core/models/generics.py
@@ -4,9 +4,8 @@ from nautobot.core.models import BaseModel
 from nautobot.core.models.fields import TagsField
 from nautobot.extras.models.change_logging import ChangeLoggedModel
 from nautobot.extras.models.customfields import CustomFieldModel
-from nautobot.extras.models.mixins import ContactMixin, DynamicGroupsModelMixin, NotesMixin
+from nautobot.extras.models.mixins import ContactMixin, DynamicGroupsModelMixin, NotesMixin, SavedViewMixin
 from nautobot.extras.models.relationships import RelationshipModel
-from nautobot.users.models import SavedViewMixin
 
 logger = logging.getLogger(__name__)
