+++ nautobot/core/models/name_color_content_types.py
@@ -12,9 +12,8 @@ from nautobot.extras.models.change_logging import ChangeLoggedModel
 # Importing CustomFieldModel, ChangeLoggedModel, RelationshipModel from  nautobot.extras.models
 # caused circular import error
 from nautobot.extras.models.customfields import CustomFieldModel
-from nautobot.extras.models.mixins import ContactMixin, DynamicGroupsModelMixin, NotesMixin
+from nautobot.extras.models.mixins import ContactMixin, DynamicGroupsModelMixin, NotesMixin, SavedViewMixin
 from nautobot.extras.models.relationships import RelationshipModel
-from nautobot.users.models import SavedViewMixin
 
 
 class ContentTypeRelatedQuerySet(RestrictedQuerySet):
