+++ nautobot/extras/models/tags.py
@@ -7,8 +7,8 @@ from nautobot.core.constants import CHARFIELD_MAX_LENGTH
 from nautobot.core.models import BaseManager, BaseModel
 from nautobot.core.models.fields import ColorField
 from nautobot.core.models.querysets import RestrictedQuerySet
+from nautobot.extras.models.mixins import SavedViewMixin
 from nautobot.extras.utils import extras_features, TaggableClassesQuery
-from nautobot.users.models import SavedViewMixin
 
 # These imports are in this particular order because of circular import problems
 from .change_logging import ChangeLoggedModel
