+++ nautobot/extras/tables.py
@@ -45,6 +45,7 @@ from .models import (
     Relationship,
     RelationshipAssociation,
     Role,
+    SavedView,
     ScheduledJob,
     Secret,
     SecretsGroup,
@@ -466,6 +467,37 @@ class NestedDynamicGroupAncestorsTable(DynamicGroupTable):
     class Meta(DynamicGroupTable.Meta):
         fields = ["name", "members", "description", "actions"]
         exclude = ["content_type"]
+
+
+class SavedViewTable(BaseTable):
+    name = tables.Column(linkify=True)
+    actions = ButtonsColumn(SavedView)
+    is_global_default = BooleanColumn()
+    is_shared = BooleanColumn()
+
+    class Meta(BaseTable.Meta):
+        model = SavedView
+        fields = (
+            "name",
+            "owner",
+            "view",
+            "config",
+            "is_global_default",
+            "is_shared",
+            "actions",
+        )
+        default_columns = (
+            "name",
+            "owner",
+            "view",
+            "is_global_default",
+            "actions",
+        )
+
+    def render_config(self, record):
+        if record.config:
+            return render_json(record.config, pretty_print=True)
+        return self.default
 
 
 class StaticGroupAssociationTable(BaseTable):
