+++ nautobot/users/api/views.py
@@ -12,7 +12,7 @@ from nautobot.core.api.views import ModelViewSet
 from nautobot.core.models.querysets import RestrictedQuerySet
 from nautobot.core.utils.data import deepmerge
 from nautobot.users import filters
-from nautobot.users.models import ObjectPermission, SavedView, Token, UserSavedViewAssociation
+from nautobot.users.models import ObjectPermission, Token
 
 from . import serializers
 
@@ -35,23 +35,6 @@ class GroupViewSet(ModelViewSet):
     serializer_class = serializers.GroupSerializer
     bulk_operation_serializer_class = BulkOperationIntegerIDSerializer
     filterset_class = filters.GroupFilterSet
-
-
-#
-# Saved Views
-#
-
-
-class SavedViewViewSet(ModelViewSet):
-    queryset = SavedView.objects.select_related("owner")
-    serializer_class = serializers.SavedViewSerializer
-    filterset_class = filters.SavedViewFilterSet
-
-
-class UserSavedViewAssociationViewSet(ModelViewSet):
-    queryset = UserSavedViewAssociation.objects.select_related("user", "saved_view")
-    serializer_class = serializers.UserSavedViewAssociationSerializer
-    filterset_class = filters.UserSavedViewAssociationFilterSet
 
 
 #
