+++ nautobot/users/factory.py
@@ -1,12 +1,9 @@
 from datetime import timezone
 
 from django.contrib.auth import get_user_model
-from django.contrib.contenttypes.models import ContentType
 import factory
 
-from nautobot.core.factory import BaseModelFactory, NautobotBoolIterator, random_instance, UniqueFaker
-from nautobot.extras.utils import FeatureQuery
-from nautobot.users.models import SavedView
+from nautobot.core.factory import BaseModelFactory, NautobotBoolIterator, UniqueFaker
 
 User = get_user_model()
 
@@ -34,17 +31,3 @@ class UserFactory(BaseModelFactory):
     last_login = factory.Faker("date_time", tzinfo=timezone.utc)
 
     # TODO config_data
-
-
-class SavedViewFactory(BaseModelFactory):
-    class Meta:
-        model = SavedView
-        exclude = ("ct",)
-
-    name = factory.LazyAttributeSequence(lambda o, n: f"Sample {o.view} Saved View - {n + 1}")
-    owner = random_instance(User, allow_null=False)
-    ct = random_instance(lambda: ContentType.objects.filter(FeatureQuery("saved_views").get_query()), allow_null=False)
-    view = factory.LazyAttribute(lambda o: f"{o.ct.app_label}:{o.ct.model}_list")
-    config = factory.Faker("pydict")
-    is_shared = NautobotBoolIterator()
-    # is_global_default currently just defaults to False for all randomly generated saved views
